SDRANGE Function
This function calculates the standard deviation for a range of numbers. The range is given by the second value. Then, the function compares the first number to the average of the range +/- the SD. Depending whether the first value is above, below, or equal/neither above/below the range average +/- the SD, the  function generates matching “high,” “low,” and “average” strings.